_old/requirements.txt
# Core Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Async Support
aiohttp>=3.10.0
asyncio-throttle>=1.0.0
